requirements.txt
jupyterlab==3.3.2
matplotlib==3.4.3
numpy==1.21.5
pandas==1.3.5
python==3.9.7
pytorch==1.11.0
cudatoolkit==11.5.1
scienceplots==1.0.9
scikit-learn==1.0.2
tensorboard==2.0.0
torchvision==0.12.0
tqdm==4.63.0